fix(checkpoints): Keep tester variables current
